Attempt at making thumbnail page fit redesign
There is no suggested redesign for the thumbnail view,
but based on feedback and the design of other view
I tried to give it a quick make-over anyway. Still not
perfect, but should be more agreeable with the redesign
and a good starting point for further discussion.
